requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp_socks==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.2.0
bottle==0.13.1
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
clr-loader==0.2.6
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1
cssselect2==0.7.0
curl_cffi==0.7.1
cycler==0.12.1
cymem==2.0.11
defusedxml==0.7.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
duckduckgo_search==6.2.12
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
g4f==0.4.6.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
loguru==0.7.2
lz4==4.3.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
plyer==2.1.0
preshed==3.0.9
primp==0.6.2
proxy_tools==0.1.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.18.0
pyparsing==3.1.4
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-socks==2.5.1
pythonnet==3.0.4
pytz==2024.2
pywebview==5.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.8.1
ruff==0.6.6
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
sqlparse==0.5.3
srsly==2.5.1
starlette==0.41.3
sympy==1.13.1
textstat==0.7.4
thinc==8.3.4
tinycss2==1.3.0
tokenizers==0.21.0
tomlkit==0.12.0
torch==2.6.0
tqdm==4.66.5
transformers==4.48.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.4
win32-setctime==1.1.0
wrapt==1.17.2
yarl==1.11.1
